unit-tests: convert t/helper/test-oid-array.c to unit-tests
Migrate t/helper/test-oid-array.c and t/t0064-oid-array.sh to the
recently added unit testing framework. This would improve runtime
performance and provide better debugging via displaying array content,
index at which the test failed etc. directly to stdout.

There is only one change in the new testing approach. In the previous
testing method, GIT_TEST_DEFAULT_HASH environment variable which is
detected via t/test-lib-functions.sh:test_detect_hash() was used to
create a new repository with hash algorithm specified by that
environment variable. And the helper binary would detect the hash
algorithm via detecting the repository type. In the new approach
however, we do not need to initialize the repo and we can just
construct the object_id according to the hash algorithm and insert it
directly into oidarray. The added benefit of this is that it becomes
easy to test multiple hash algorithm being used on a single oidarray in
the future.

This also prints the oidarray and the input/expected result on stdout
when a particular test fails, so that it becomes easy to debug.

Signed-off-by: Ghanshyam Thakkar <shyamthakkar001@gmail.com>