setup: remove unnecessary variable
The TODO comment suggests to heed core.bare from template config file if
no command line override given. And the prev_bare_repository variable
seems to have been placed for this sole purpose as it is not used
anywhere else.

However, it was clarified by Junio [1] that such values are ignored
intentionally and does not make sense to propagate it from template
config to repository config. Also, the directories for the worktree and
repository are already created by the point we reach the codepath where
the TODO comment is placed. Therefore, prev_bare_repository does not
have a usecase with/without supporting core.bare from template. And the
removal of prev_bare_repository is safe as proved by the later part of
the comment:

    "Unfortunately, the line above is equivalent to
        is_bare_repository_cfg = !work_tree;
    which ignores the config entirely even if no `--[no-]bare`
    command line option was present.
    To see why, note that before this function, there was this call:
        prev_bare_repository = is_bare_repository()
    expanding the right hand side:
        = is_bare_repository_cfg && !get_git_work_tree()
        = is_bare_repository_cfg && !work_tree
    note that the last simplification above is valid because nothing
    calls repo_init() or set_git_work_tree() between any of the
    relevant calls in the code, and thus the !get_git_work_tree()
    calls will return the same result each time.  So, what we are
    interested in computing is the right hand side of the line of
    code just above this comment:
        prev_bare_repository || !work_tree
        = is_bare_repository_cfg && !work_tree || !work_tree
        = !work_tree
    because "A && !B || !B == !B" for all boolean values of A & B."

Therefore, remove the TODO comment and remove prev_bare_repository
variable. Also, update relevant testcases and remove one redundant
testcase.

[1]: https://lore.kernel.org/git/xmqqjzonpy9l.fsf@gitster.g/

Helped-by: Elijah Newren <newren@gmail.com>
Helped-by: Junio C Hamano <gitster@pobox.com>
Signed-off-by: Ghanshyam Thakkar <shyamthakkar001@gmail.com>