builtin/commit: error out when passing untracked path with -i
Currently when we provide a pathspec which does not match any tracked
path alongside --include, we do not error like without --include. If
there something staged, it will commit the staged changes and ignore the
pathspec which does not match any tracked path. And if nothing is
staged, it will print the status. Exit code is 0 in both cases (unlike
without --include). This described in the TODO comment before the
relevant testcase.

Fix this by matching the pathspec against index and report error if
any. And amend the relevant testcase and remove the TODO comment.

Signed-off-by: Ghanshyam Thakkar <shyamthakkar001@gmail.com>