builtin/add: error out when passing untracked path with -u
When passing untracked path with -u option, it silently succeeds.
There is no error message and the exit code is zero. This is
inconsistent with other instances of git commands where the expected
argument is a known path. In those other instances, we error out when
the path is not known.

Therefore, fix this by passing a character array to
add_files_to_cache() to collect the pathspec matching information and
report the error if a pathspec does not match any cache entry. Also add
a testcase to cover this scenario.

Signed-off-by: Ghanshyam Thakkar <shyamthakkar001@gmail.com>