t/: port helper/test-sha1.c to unit-tests/t-hash.c
t/helper/test-sha1 and t/t0015-hash.sh test the basic hash
implementation of sha1 in Git. Migrate them to the new unit testing
framework for better debugging and simplicity.

The sha1 subcommand from test-tool is still not removed because it is
relied upon by t0013-sha1dc (which requires 'test-tool sha1' dying when
it is used on a file created to contain the known sha1 attack.) and
pack_trailer():lib-pack.sh

Helped-by: Patrick Steinhardt <ps@pks.im>
Mentored-by: Christian Couder <chriscool@tuxfamily.org>
Mentored-by: Kaartic Sivaraam <kaartic.sivaraam@gmail.com>
Co-authored-by: Achu Luma <ach.lumap@gmail.com>
Signed-off-by: Achu Luma <ach.lumap@gmail.com>
Signed-off-by: Ghanshyam Thakkar <shyamthakkar001@gmail.com>