t/: port helper/test-sha256.c to unit-tests/t-hash.c
t/helper/test-sha256 and t/t0015-hash test the basic hash implementation
of sha256 in Git. Port them to the new unit testing framework for better
debugging, simplicity and faster runtime. The necessary building blocks
are already implemented in t-hash in a previous commit which ported
test-sha1.

The 'sha256' subcommand of test-tool is still not removed, because it
is used by pack_trailer() in lib-pack.sh, which is used in many tests
of the t53** series.

Helped-by: Patrick Steinhardt <ps@pks.im>
Mentored-by: Christian Couder <chriscool@tuxfamily.org>
Mentored-by: Kaartic Sivaraam <kaartic.sivaraam@gmail.com>
Co-authored-by: Achu Luma <ach.lumap@gmail.com>
Signed-off-by: Achu Luma <ach.lumap@gmail.com>
Signed-off-by: Ghanshyam Thakkar <shyamthakkar001@gmail.com>