t: migrate helper/test-urlmatch-normalization to unit tests
helper/test-urlmatch-normalization along with
t0110-urlmatch-normalization test the `url_normalize()` function from
'urlmatch.h'. Migrate them to the unit testing framework for better
performance and debugging.

In the migration, last two of the checks from `t_url_general_escape()`
were slightly changed. This involved changing the '\' character to
'\\', since only '\' will act as an escape character, but we want the
'\' itself. And using the '%5C' hex identifier to check for the single
'\'.

Mentored-by: Christian Couder <chriscool@tuxfamily.org>
Mentored-by: Kaartic Sivaraam <kaartic.sivaraam@gmail.com>
Signed-off-by: Ghanshyam Thakkar <shyamthakkar001@gmail.com>